% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_rothc.R
\name{ag_rothC}
\alias{ag_rothC}
\title{Agrenna RothC full implementation}
\usage{
ag_rothC(x, write = F, cf1 = 1, cf2 = 1, cf3 = 1, cf4 = 1, cf5 = 1, cf6 = 1)
}
\arguments{
\item{x}{google sheets input}

\item{write}{write output of management, only used for debugging!}

\item{cf1}{Slope yield inputs}

\item{cf2}{Intercept yield inputs}

\item{cf3}{DR ratio}

\item{cf4}{Decomposition rate}

\item{cf5}{Evapotranspiration coeffj}
}
\description{
Process and run RothC.
}
\author{
Marcos Alves
}
