% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CFGs}
\alias{CFGs}
\title{This is data to be included in my package}
\format{
An object of class \code{data.frame} with 12 rows and 2 columns.
}
\usage{
CFGs
}
\description{
This is data to be included in my package
}
\author{
Marcos Alves
}
\keyword{data}
