% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_add_columns.R
\name{check_and_add_columns}
\alias{check_and_add_columns}
\title{Check and Add Missing Columns}
\usage{
check_and_add_columns(df, required_columns)
}
\arguments{
\item{df}{The input data.frame.}
}
\value{
The modified data.frame with the missing columns added.
}
\description{
This function takes a data.frame as input and checks if the specified columns are present. If any of the columns are missing, they are added to the data.frame in the specified order.
}
\examples{
df <- data.frame(year = 2022, ini_soc = 1, Interval_1 = 0, Interval_2 = 0)
df_modified <- check_and_add_columns(df)
str(df_modified)
}
