% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_crop_convertion.R
\name{cover_crop_convertion}
\alias{cover_crop_convertion}
\title{Cover crop name convertion}
\usage{
cover_crop_convertion(ccname)
}
\arguments{
\item{ccname}{vector with cropnames}
}
\value{
Vector with monthly values
}
\description{
convert a common name to a list of logical values
}
\author{
Marcos Alves
}
