% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsw}
\alias{dsw}
\title{This is data to be included in my package}
\format{
An object of class \code{list} of length 11.
}
\usage{
dsw
}
\description{
This is data to be included in my package
}
\author{
Marcos Alves
}
\keyword{data}
