% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmental_variables.R
\name{environmental_variables}
\alias{environmental_variables}
\title{Environmental Variables}
\usage{
environmental_variables(lat, long, soil_depth)
}
\arguments{
\item{lat}{Latitude}

\item{long}{Longitute}

\item{soil_depth}{Soil depth}
}
\description{
Download and preprocess environmental variables.
}
\author{
Marcos Alves
}
