% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_data_set.R
\name{expand_data_set}
\alias{expand_data_set}
\title{Expand google sheets datasets}
\usage{
expand_data_set(x)
}
\arguments{
\item{x}{googlesheet dataset}
}
\description{
expand nested soil carbon values in sequential rows
}
\author{
Marcos Alves
}
