% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_vcu_ledger.R
\name{generate_vcu_ledger}
\alias{generate_vcu_ledger}
\title{Calculate and Round VCUs}
\usage{
generate_vcu_ledger(full_cals, file_name)
}
\arguments{
\item{full_cals}{A data frame containing full calculations.}

\item{file_name}{Character string specifying the output file name for the results.}
}
\value{
A data frame containing calculated and rounded VCUs, and the results are written to an Excel file.
}
\description{
This function calculates and rounds Verified Carbon Units (VCUs) based on a set of input parameters.
}
\examples{
# Example usage:
generate_vcu_ledger(full_cals = my_data, file_name = "output.xlsx")

}
