% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gee_isric_soil_profile_rothc.R
\name{get_gee_isric_soil_profile_rothc}
\alias{get_gee_isric_soil_profile_rothc}
\title{Extract soil profile data from ISRIC SoilGrids using Google Earth Engine.}
\usage{
get_gee_isric_soil_profile_rothc(df_latlon)
}
\arguments{
\item{df_latlon}{A data frame containing latitude and longitude coordinates.}
}
\value{
A list of data frames containing extracted soil profile data.
}
\description{
This function extracts soil profile data including Organic Carbon Stock (OCS),
Bulk Density (BD), Clay content, Sand content, and Silt content from ISRIC SoilGrids
using Google Earth Engine.
}
\examples{
df_latlon <- data.frame(lon = c(-75.5, -76.0), lat = c(42.0, 42.5))
soil_data <- get_gee_isric_soil_profile_rothc(df_latlon)

}
