% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ini_carbon_pools.R
\name{ini_carbon_pools}
\alias{ini_carbon_pools}
\title{Initialize soil carbon}
\usage{
ini_carbon_pools(ini_carbon, env_clay, xi, cf4)
}
\arguments{
\item{ini_carbon}{initial C value}

\item{env_clay}{clay content}

\item{xi}{decomposition rate factors}

\item{cf4}{Calibration factors for Decomposition rates}
}
\description{
Find the equilibrium initial soil carbon for all rothc carbon pools.
}
\author{
Marcos Alves
}
