% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lucas}
\alias{lucas}
\title{This is data to be included in my package}
\format{
An object of class \code{data.frame} with 18984 rows and 39 columns.
}
\usage{
lucas
}
\description{
This is data to be included in my package
}
\author{
Marcos Alves
}
\keyword{data}
