% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management_variables.R
\name{management_variables}
\alias{management_variables}
\title{Environmental Variables}
\usage{
management_variables(
  mgmt_file,
  flag = c("Treatment", "Control"),
  write = FALSE,
  cf1 = 1,
  cf2 = 1,
  cf3 = 1
)
}
\arguments{
\item{mgmt_file}{google sheets with field data}

\item{flag}{separete treatment from control fields}

\item{write}{write output of management, only used for debugging!}

\item{cf1}{Coefficient factor for slope yield inputs}

\item{cf2}{Coefficient factor Intercept yield inputs}

\item{cf3}{Coefficient factor DR ratio}
}
\description{
Download and preprocess management variables.
}
\author{
Marcos Alves
}
