% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{post_processing}
\alias{post_processing}
\title{Post process output from \link{AgreenaProgramme2}}
\usage{
post_processing(out, type = "payout", conversion = 44/12)
}
\arguments{
\item{out}{Object returned from \link{AgreenaProgramm2}.}

\item{type}{String: Currently ununsed.}

\item{conversion}{Numeric: Conversion factor for C =>-CO2e conversion.}
}
\description{
Note that, by default, this converts RothC output values from tC
to \strong{tCO2e}.
}
