% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_field_data_fert.R
\name{pre_field_data_fert}
\alias{pre_field_data_fert}
\title{Process Field and Fertilizer Data}
\usage{
pre_field_data_fert(actuals_year, N_limit = 300)
}
\arguments{
\item{actuals_year}{The year for which the data should be processed.}

\item{N_limit}{The nitrogen limit for scaling application rates (default is 300).}
}
\value{
A dataframe containing processed field and fertilizer data.
}
\description{
This function reads field and fertilizer data from CSV or Excel files, calculates
the nitrogen and carbon contributions from fertilizers, and joins the data into
a single dataframe.
}
\details{
This function performs the following steps:
\enumerate{
\item Determines the file delimiter for CSV files.
\item Reads field and fertilizer data from specified files.
\item Converts character columns in field_data to numeric if possible.
\item Reads organic carbon content rates from a CSV file.
\item Filters and selects relevant columns from the fertilizer data.
\item Computes nitrogen rates based on fertilizer data and organic nitrogen content.
\item Scales application rates to stay within the nitrogen limit (N_limit).
\item Computes carbon rates based on scaled application rates.
\item Joins the processed data into a single dataframe.
}
}
\examples{
\dontrun{
  pre_field_data_fert(2022)
}

}
