% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_fractions}
\alias{res_fractions}
\title{This is data to be included in my package}
\format{
An object of class \code{data.frame} with 26 rows and 12 columns.
}
\usage{
res_fractions
}
\description{
This is data to be included in my package
}
\author{
Marcos Alves
}
\keyword{data}
