% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resid_mgmt}
\alias{resid_mgmt}
\title{This is data to be included in my package}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 12 columns.
}
\usage{
resid_mgmt
}
\description{
This is data to be included in my package
}
\author{
Marcos Alves
}
\keyword{data}
