% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residue_fractions.R
\name{residue_fractions}
\alias{residue_fractions}
\title{Fraction of residues}
\usage{
residue_fractions(cropname)
}
\arguments{
\item{cropname}{vector with cropnames}
}
\value{
Vector with monthly values
}
\description{
Select the amount of the residues being left on the field in each month
}
\author{
Marcos Alves
}
