% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residue_management.R
\name{residue_management}
\alias{residue_management}
\title{Fraction of residues}
\usage{
residue_management(res_mgmt)
}
\arguments{
\item{res_mgmt}{vector management options}
}
\value{
Vector with monthly values
}
\description{
Select the amount of the residues being left on the field in each month
}
\author{
Marcos Alves
}
