% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tillage_convert}
\alias{tillage_convert}
\title{This is data to be included in my package}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 1 columns.
}
\usage{
tillage_convert
}
\description{
This is data to be included in my package
}
\author{
Marcos Alves
}
\keyword{data}
