% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tillage_convertion.R
\name{tillage_convertion}
\alias{tillage_convertion}
\title{Tillage convertion factor}
\usage{
tillage_convertion(tillage)
}
\arguments{
\item{tillage}{vector with tillage regimes}
}
\value{
Vector with monthly values
}
\description{
Select the convertion factor depending on the tillage regime
}
\author{
Marcos Alves
}
