% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ini_soc.R
\name{update_ini_soc}
\alias{update_ini_soc}
\title{Update ini_soc column based on field_id and actual_harvest_year}
\usage{
update_ini_soc(data)
}
\arguments{
\item{data}{The input dataset}
}
\value{
The updated dataset with the ini_soc column modified
}
\description{
This function checks the number of occurrences of each field_id in the dataset.
For field_ids that occur only once, it assigns the value "isric" to the ini_soc column.
For field_ids with multiple rows, it assigns the value of the previous year's interval_1 to the ini_soc column.
}
