% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_previous_hy_soc.R
\name{update_previous_hy_soc}
\alias{update_previous_hy_soc}
\title{Replace values in a data frame with values from a CSV file}
\usage{
update_previous_hy_soc(csv_file, df)
}
\arguments{
\item{csv_file}{The path to the CSV file.}

\item{df}{The original data frame.}
}
\value{
A data frame with replaced values.
}
\description{
This function replaces the values in the original data frame with the values from a CSV file
for the rows where 'field_id' and 'actual_harvest_year' are the same and the column names are the same in both data frames.
}
\examples{
\dontrun{
replace_values("path/to/your/csv_file.csv", your_data_frame)
}
}
