% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_out.R
\name{write_out}
\alias{write_out}
\title{Write out data to CSV files with CO2 SOC removals information.}
\usage{
write_out(out, field_data, sim_period)
}
\arguments{
\item{out}{A data frame containing the input data, including columns 'field_id', 'year', and 'co2_removals'.
'field_id' uniquely identifies each field, 'year' represents the year of observation,
and 'co2_removals' contains the recorded CO2 SOC removals for that year and field.}

\item{field_data}{A data frame containing additional information about the fields,
including 'field_id' and 'field_size_ha', which indicates the size of each field in hectares.
This information is used to calculate the total CO2 SOC removals per field.}
}
\value{
This function doesn't return anything directly, but it writes the processed data to CSV files.
}
\description{
This function takes a data frame 'out', which contains information about CO2
SOC (Soil Organic Carbon) removals for various fields over multiple years.
It performs additional data transformations and writes the results to multiple CSV files:
}
\details{
\enumerate{
\item A CSV file containing CO2 SOC removals per year per field during the project duration.
\item A CSV file containing the total CO2 SOC removals per year across all fields during the project duration.
\item A CSV file containing CO2 SOC removals per field for the harvest year (and a copy in .rds)
}
}
\examples{
# Example usage:
# write_out(out_data, field_data)
}
