% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xi_calc.R
\name{xi_calc}
\alias{xi_calc}
\title{AgRothC}
\usage{
xi_calc(env, mgmt, cf5 = 1)
}
\arguments{
\item{env}{Environmental pre processed data}

\item{mgmt}{management preprocessed data}

\item{cf5}{calibration factor Evapotranspiration coeff}
}
\description{
Process and run RothC.
}
\author{
Marcos Alves
}
