% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yld2bio}
\alias{yld2bio}
\title{This is data to be included in my package}
\format{
An object of class \code{data.frame} with 26 rows and 4 columns.
}
\usage{
yld2bio
}
\description{
This is data to be included in my package
}
\author{
Marcos Alves
}
\keyword{data}
